Frontend: added User Hooks
